+++ src/config.js
@@ -3,12 +3,14 @@ import program from 'commander';
 import defaultOptions from './default-options.json';
 import pkg from '../package.json';
 
+
 const availableOptions = [
     [ 'm', 'automock', 'mock every module in the project automatically' ],
     [ 'j', 'jsdom', 'use jsdom' ],
     [ 'w', 'watch', 'watch for file changes' ],
     [ 'c', 'coverage', 'generate a code coverage report' ],
-    [ null, 'no-babel', 'disable babel support' ],
+    [ 'd', 'test-directory <name>', 'test directory name' ],
+    [ 'B', 'no-babel', 'disable babel support' ],
 ];
 
 
@@ -36,13 +38,15 @@ try {
 
 const parsedOptions = {};
 
-
 for (var i = 0; i < availableOptions.length; i++) {
-    const key = availableOptions[i][1];
+    const key = availableOptions[i][1].split(' ')[0].replace(/-./, s => s[1].toUpperCase());
     if (program.hasOwnProperty(key)) {
         parsedOptions[key] = program[key];
     }
 }
 
+if (program.babel === false) {
+    parsedOptions.babel = false;
+}
 
 export const options = Object.assign(defaultOptions, rc, parsedOptions);