+++ src/init-test.js
@@ -1,7 +1,5 @@
 import { options } from './config';
-import { cleanUp as cleanUpAutomock, reset as resetAutomock } from './automock';
-import { cleanUp as cleanUpMokamok } from './mokamok';
-import { cleanUp as cleanUpRequireHook } from './require-hook';
+
 
 let jsDomCleanUp = null;
 
@@ -14,15 +12,7 @@ before(function () {
 
 
 afterEach(function () {
-    resetAutomock();
-});
-
-
-after(function () {
     if (options.jsdom) {
         jsDomCleanUp();
     }
-    cleanUpMokamok();
-    cleanUpAutomock();
-    cleanUpRequireHook();
 });