+++ src/run-all-tests.js
@@ -1,29 +1,26 @@
 import glob from 'glob';
-import runTest from './run-test';
+import { addTestFile, runTests } from './run-test';
+import { options } from './config';
 
 export default function runAllTests(callback) {
     return new Promise((resolve, reject) => {
-        glob("**/--tests--/*.js", function (err, files) {
+        glob(`**/${options.testDirectory}/*.js`, function (err, files) {
             if (err) {
                 console.log(err);
                 process.exit(1);
             }
             files = files.filter(fname => fname.indexOf('node_modules') === -1);
-            function next(idx) {
-                if (idx < files.length) {
-                    runTest(files[idx]).then(err => {
-                        if (err) {
-                            reject(err);
-                        }
-                        next(idx + 1);
-                    }).catch((err) => {
-                        reject(err);
-                    });
-                } else {
-                    resolve();
-                }
+            for (var i = 0; i < files.length; i++) {
+                addTestFile(files[i]);
             }
-            next(0);
+            runTests().then(err => {
+                if (err) {
+                    reject(err);
+                }
+                resolve();
+            }).catch((err) => {
+                reject(err);
+            });
         });
     });
 };