+++ tests/code-coverage/test.js
@@ -1,0 +1,7 @@
+export function fn1() {
+    return 1;
+}
+
+export function fn2() {
+    return 2;
+}