+++ tests/jsdom/--tests--/spec.js
@@ -1,7 +1,7 @@
 describe("Jsdom support", () => {
 
     it("init jsdom", () => {
-        expect(window).to.not.equal(null);
+        expect(!!window).to.equal(true);
     });
 
 });