+++ tests/jsdom/config.js
@@ -1,0 +1,19 @@
+module.exports = {
+
+    tests: [
+        {
+            name: 'with-long-parameter',
+            params: '--jsdom'
+        },
+        {
+            name: 'with-short-parameter',
+            params: '-j'
+        },
+        {
+            rc: {
+                jsdom: true
+            }
+        }
+    ]
+
+}